+++ app/Scraper/Manager.php
@@ -20,6 +20,7 @@ final class Manager
         \Kami\Cocktail\Scraper\Sites\CocktailParty::class,
         \Kami\Cocktail\Scraper\Sites\LiberAndCo::class,
         \Kami\Cocktail\Scraper\Sites\CocktailsDistilled::class,
+        \Kami\Cocktail\Scraper\Sites\PunchDrink::class,
     ];
 
     public function __construct(private readonly string $url)